Move the URI type definition from custom_data_types to a new module called data_types, in mytardis_client. The reasoning is that the format of the URI is closely coupled to MyTardis itself.
