+++ src/ingestion_factory/factory.py
@@ -5,7 +5,6 @@ import logging
 from pathlib import Path
 from typing import Any, Optional
 
-from src.blueprints.custom_data_types import URI
 from src.blueprints.datafile import Datafile, RawDatafile
 from src.blueprints.dataset import RawDataset
 from src.blueprints.experiment import RawExperiment
@@ -15,6 +14,7 @@ from src.conveyor.conveyor import Conveyor, FailedTransferException
 from src.crucible.crucible import Crucible
 from src.extraction.manifest import IngestionManifest
 from src.forges.forge import Forge
+from src.mytardis_client.data_types import URI
 from src.mytardis_client.mt_rest import MyTardisRESTFactory
 from src.mytardis_client.objects import MyTardisObject
 from src.overseers.overseer import Overseer
