+++ src/overseers/overseer.py
@@ -10,8 +10,8 @@ from urllib.parse import urljoin
 from pydantic import ValidationError
 from requests.exceptions import HTTPError
 
-from src.blueprints.custom_data_types import URI
 from src.config.config import IntrospectionConfig
+from src.mytardis_client.data_types import URI
 from src.mytardis_client.endpoints import get_endpoint
 from src.mytardis_client.mt_rest import MyTardisRESTFactory
 from src.mytardis_client.objects import MyTardisObject, get_type_info
