+++ tests/test_dataclasses.py
@@ -5,7 +5,8 @@
 from pytest import fixture
 
 from src.blueprints.common_models import Parameter, ParameterSet
-from src.blueprints.custom_data_types import URI, MTUrl
+from src.blueprints.custom_data_types import MTUrl
+from src.mytardis_client.data_types import URI
 
 
 @fixture
